Simplified directory tree by moving everything up two levels.
