+++ Encryptamajig.Tests/EncryptamajigTests.cs
