+++ Encryptamajig.Tests/Properties/AssemblyInfo.cs
