+++ Encryptamajig.sln
@@ -1,0 +1,60 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Encryptamajig", "Encryptamajig\Encryptamajig.csproj", "{C9385513-156E-43BB-9120-110905C5C528}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Encryptamajig.Tests", "Encryptamajig.Tests\Encryptamajig.Tests.csproj", "{7416C94A-18CA-42F6-BD4F-1F113A8BC303}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{9CC0FE37-15F0-4FD1-8FA2-E3A4D5A2E8DE}"
+	ProjectSection(SolutionItems) = preProject
+		.nuget\NuGet.exe = .nuget\NuGet.exe
+		.nuget\NuGet.targets = .nuget\NuGet.targets
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Encryptamajig.Tester", "Encryptamajig.Tester\Encryptamajig.Tester.csproj", "{9E26B709-E8E8-4D23-A3D5-FF6851F78751}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C9385513-156E-43BB-9120-110905C5C528}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{C9385513-156E-43BB-9120-110905C5C528}.Release|x86.ActiveCfg = Release|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{7416C94A-18CA-42F6-BD4F-1F113A8BC303}.Release|x86.ActiveCfg = Release|Any CPU
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Debug|Mixed Platforms.Build.0 = Debug|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Debug|x86.ActiveCfg = Debug|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Debug|x86.Build.0 = Debug|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Release|Any CPU.ActiveCfg = Release|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Release|Mixed Platforms.ActiveCfg = Release|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Release|Mixed Platforms.Build.0 = Release|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Release|x86.ActiveCfg = Release|x86
+		{9E26B709-E8E8-4D23-A3D5-FF6851F78751}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
