+++ Encryptamajig/Properties/AssemblyInfo.cs
